changes for comments/dashboard access
