+++ met-web/src/components/engagement/dashboard/comment/CommentsBlock.tsx
@@ -1,12 +1,45 @@
 import React, { useContext } from 'react';
 import { Grid, Skeleton } from '@mui/material';
-import { Link } from 'react-router-dom';
+import { Link, useNavigate } from 'react-router-dom';
 import { CommentViewContext } from './CommentViewContext';
 import { PrimaryButton, MetPaper, MetHeader4 } from 'components/common';
 import CommentTable from './CommentTable';
+import { useAppSelector, useAppDispatch } from 'hooks';
+import { SubmissionStatus } from 'constants/engagementStatus';
+import { openNotificationModal } from 'services/notificationModalService/notificationModalSlice';
 
 export const CommentsBlock = () => {
     const { engagement, isEngagementLoading } = useContext(CommentViewContext);
+    const navigate = useNavigate();
+    const dispatch = useAppDispatch();
+    const roles = useAppSelector((state) => state.user.roles);
+
+    const handleViewDashboard = () => {
+        if (roles.length !== 0) {
+            navigate(`/engagements/${engagement?.id}/dashboard`);
+            return;
+        }
+
+        if (engagement?.submission_status === SubmissionStatus.Closed) {
+            navigate(`/engagements/${engagement.id}/dashboard`);
+            return;
+        }
+
+        dispatch(
+            openNotificationModal({
+                open: true,
+                data: {
+                    header: 'View Report',
+                    subText: [
+                        {
+                            text: 'The report will only be available to view after the engagement period is over and the engagement is closed.',
+                        },
+                    ],
+                },
+                type: 'update',
+            }),
+        );
+    };
 
     if (isEngagementLoading || !engagement) {
         return <Skeleton width="100%" height="40em" />;
@@ -35,8 +68,7 @@ export const CommentsBlock = () => {
                         >
                             <PrimaryButton
                                 data-testid="SurveyBlock/take-me-to-survey-button"
-                                component={Link}
-                                to={`/engagements/${engagement.id}/dashboard`}
+                                onClick={handleViewDashboard}
                             >
                                 View Report
                             </PrimaryButton>
